digiplus_hr/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
Pillow==12.0.0
django-filter==24.3
python-decouple==3.8
channels 
channels-redis
redis
websockets
daphne==4.1.0
whitenoise==6.8.1
psycopg2-binary
gunicorn
dj-database-url==2.1.0
qrcode==7.4.2
drf-spectacular
reportlab==4.1.0

# IA / DeepSeek / RAG dependencies
requests
openai
langchain
langchain-community
langchain-text-splitters
sentence-transformers
faiss-cpu
PyPDF2
